+++ packages/on-chain-relayer/contracts/custom-integrations/mento/linkedlists/AddressSortedLinkedListWithMedian.sol
@@ -18,11 +18,11 @@ library AddressSortedLinkedListWithMedian {
   using SortedLinkedListWithMedian for SortedLinkedListWithMedian.List;
 
   function toBytes(address a) public pure returns (bytes32) {
-    return bytes32(uint256(a) << 96);
+    return bytes32(uint256(uint160(a)) << 96);
   }
 
   function toAddress(bytes32 b) public pure returns (address) {
-    return address(uint256(b) >> 96);
+    return address(uint160(uint256(b) >> 96));
   }
 
   /**
